flatten tasks for celery autodiscovery
