+++ deploy/compose/docker-compose.yml
@@ -37,26 +37,6 @@ services:
     environment:
       - DEBUG=
 
-  worker:
-    build: ../../backend
-    command: "celery -A app worker -l info"
-    env_file:
-      - ./.env
-    depends_on:
-      - backend
-    environment:
-      - DEBUG=
-
-  scheduler:
-    build: ../../backend
-    command: "celery -A app beat -l info"
-    env_file:
-      - ./.env
-    depends_on: 
-      - backend
-    environment: 
-      - DEBUG=
-
   frontend:
     build:
       context: ../../frontend
