+++ deploy/staging/deploy.sh
@@ -62,7 +62,7 @@ terraform --version;
 configureAWS;
 pushFrontendToECR;
 pushBackendToECR;
-pushCompressionServiceToECR;
+# pushCompressionServiceToECR;
 createInfrastructure;
 createPlatform;
 createDatabase;
